Isolate generic code in separate file
