+++ external/vulkancts/modules/vulkan/metamorphic/vktMetamorphic.hpp
@@ -1,0 +1,39 @@
+/*-------------------------------------------------------------------------
+ * Vulkan Conformance Tests
+ * ------------------------
+ *
+ * Copyright (c) 2018 Google Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ *//*!
+ * \file
+ * \brief Metamorphic Tests
+ *//*--------------------------------------------------------------------*/
+
+#include "vktTestCase.hpp"
+
+namespace vkt
+{
+namespace metamorphic
+{
+
+typedef struct UniformEntry {
+  size_t size;
+  void *value;
+} UniformEntry;
+
+tcu::TestStatus renderShaderPair (Context& context, std::vector<UniformEntry> uniformEntries);
+
+} // metamorphic
+} // vkt
